requirements.txt
python-telegram-bot==11.1.0
redis==4.3.4
requests==2.28.1
python-slugify==6.1.*
geopy==2.2.*